server/ml-predict/requirements.txt
numpy==2.3.4
pillow==12.0.0
torch==2.9.0+cpu
torchvision==0.24.0+cpu
colorthief==0.2.1
boto3==1.34.0
PyJWT==2.8.0
--extra-index-url https://download.pytorch.org/whl/cpu